+++ core/utils/misc/semgrep_scan.py
